Added code for Cookie module
We use JavaScript cookies to track data of articles the user has opened on the website. Separate functions are defined for creating and retrieving cookies. A function is also defined to detect when a page opens and extract the article type and article title from the URL of the page, and then create a cookie and store this information in it. An object is created and stored in the cookie to keep track of how many of each type of article the user has read.